pyproject.toml
[tool.poetry]
name = "etp"
version = "0.3.0"
description = "Ethiopian Identity Provider platform"
authors = ["Mohamed Nesredin <m.n.kaizen@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "example.com"
documentation = ""
repository = "https://github.com/Mohamed-Kaizen/ethiopian-identity-provider/"
keywords = ["etp"]

[tool.poetry.dependencies]
python = "3.8.4"
argon2-cffi = "^20.1.0"
confusable_homoglyphs = "^3.2.0"
dj-database-url = "^0.5.0"
django = "^3.0.8"
djangorestframework = "^3.11.0"
django-axes = "^5.4.1"
django-cors-headers = "^3.4.0"
django-filter = "^2.3.0"
django-import-export = "2.2"
django-storages = {extras = ["dropbox"], version = "^1.9.1"}
drf-yasg = "^1.17.1"
gunicorn = "^20.0.4"
Pillow = "^7.2.0"
psycopg2-binary = "^2.8.5"
python-decouple = "^3.3"
pwned-passwords-django = "^1.4"
whitenoise = {extras = ["brotli"], version = "^5.1.0"}
django-oauth-toolkit = "^1.3.2"
django-reversion = "^3.0.7"
django-reversion-compare = "^0.12.2"
django-countries = "^6.1.2"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
coverage = {extras = ["toml"], version = "^5.2"}
django-coverage-plugin = "^1.8.0"
django-debug-toolbar = "^2.2"
darglint = "^1.5.1"
flake8 = "^3.8.3"
flake8-annotations = "^2.3.0"
flake8-bandit = "^2.1.2"
flake8-black = "^0.2.0"
flake8-bugbear = "^20.1.4"
flake8-docstrings = "^1.5.0"
flake8-import-order = "^0.18.1"
isort = {extras = ["pyproject"], version = "^5.1"}
mypy = "^0.782"
mkdocs = "^1.1.2"
mkdocs-material = "^5.5"
mkdocstrings = "^0.12.1"
mkdocs-minify-plugin = "^0.3.0"
mkdocs-git-revision-date-localized-plugin = "^0.7"
pytest = "^5.4.3"
pytest-cov = "^2.10.0"
pytest-django = "^3.9.0"
radon = "^4.1.0"
safety = "^1.9.0"
xdoctest = "^0.13.0"
django-extensions = "^3.0.4"
pygraphviz = "^1.5"

[tool.coverage.run]
branch = true
source = ["etp"]
plugins = "django_coverage_plugin"

[tool.coverage.report]
show_missing = true
fail_under = 84

[tool.isort]
known_typing = "typing,types,typing_extensions,mypy,mypy_extensions"
sections = "FUTURE,TYPING,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
include_trailing_comma = true
default_section = "FIRSTPARTY"
multi_line_output = 3
force_grid_wrap = 0
use_parentheses = true
line_length = 80

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
